+++ framework/wazuh/agent.py
@@ -8,6 +8,7 @@ from wazuh.exception import WazuhException
 from wazuh.ossec_queue import OssecQueue
 from wazuh.ossec_socket import OssecSocket
 from wazuh.database import Connection
+from wazuh.InputValidator import InputValidator
 from wazuh import manager
 from wazuh import common
 from glob import glob
@@ -22,6 +23,7 @@ from pwd import getpwnam
 from grp import getgrnam
 from time import time, sleep
 import socket
+import hashlib
 from distutils.version import StrictVersion
 from itertools import groupby, chain
 try:
@@ -31,7 +33,7 @@ except ImportError:
 
 def create_exception_dic(id, e):
     """
-    Creates a dictionary with a list of agent ids and it's error codes. 
+    Creates a dictionary with a list of agent ids and it's error codes.
     """
     exception_dic = {}
     exception_dic['id'] = id
@@ -441,7 +443,6 @@ class Agent:
         :param force: Remove old agents with same IP if disconnected since <force> seconds
         :return: Agent ID.
         """
-
         manager_status = manager.status()
         if 'ossec-authd' not in manager_status or manager_status['ossec-authd'] != 'running':
             data = self._add_manual(name, ip, id, key, force)
@@ -515,6 +516,18 @@ class Agent:
             raise WazuhException(1709)
 
         force = force if type(force) == int else int(force)
+
+        # Check manager name
+        db_global = glob(common.database_path_global)
+        if not db_global:
+            raise WazuhException(1600)
+
+        conn = Connection(db_global[0])
+        conn.execute("SELECT name FROM agent WHERE (id = 0)")
+        manager_name = str(conn.fetch()[0])
+
+        if name == manager_name:
+            raise WazuhException(1705, name)
 
         # Check if ip, name or id exist in client.keys
         last_id = 0
@@ -593,7 +606,7 @@ class Agent:
 
         self.id = agent_id
 
-
+        
     def get_agent_attr(self, attr):
         """
         Returns a string with an agent's os name
@@ -721,6 +734,39 @@ class Agent:
 
         return data
 
+      
+    def _remove_single_group(self, group_id):
+        """
+        Remove the group in every agent.
+
+        :param group_id: Group ID.
+        :return: Confirmation message.
+        """
+
+        if group_id.lower() == "default":
+            raise WazuhException(1712)
+
+        if not self.group_exists(group_id):
+            raise WazuhException(1710, group_id)
+
+        ids = []
+
+        # Remove agent group
+        agents = self.get_agent_group(group_id=group_id, limit=None)
+        for agent in agents['items']:
+            self.unset_group(agent['id'])
+            ids.append(agent['id'])
+
+        # Remove group directory
+        group_path = "{0}/{1}".format(common.shared_path, group_id)
+        group_backup = "{0}/groups/{1}_{2}".format(common.backup_path, group_id, int(time()))
+        if path.exists(group_path):
+            move(group_path, group_backup)
+
+        # msg = "Group '{0}' removed.".format(group_id)
+
+        # return {'msg': msg, 'affected_agents': ids}
+
 
     @staticmethod
     def get_agents_overview(status="all", os_platform="all", os_version="all", offset=0, limit=common.database_limit, sort=None, search=None):
@@ -811,10 +857,10 @@ class Agent:
             query += ' ORDER BY id ASC'
 
 
-
-        query += ' LIMIT :offset,:limit'
-        request['offset'] = offset
-        request['limit'] = limit
+        if limit:
+            query += ' LIMIT :offset,:limit'
+            request['offset'] = offset
+            request['limit'] = limit
 
         conn.execute(query.format(','.join(select)), request)
 
@@ -1186,7 +1232,7 @@ class Agent:
         return data
 
     @staticmethod
-    def get_all_groups(offset=0, limit=common.database_limit, sort=None, search=None):
+    def get_all_groups(offset=0, limit=common.database_limit, sort=None, search=None, hash_algorithm='md5'):
         """
         Gets the existing groups.
 
@@ -1196,11 +1242,24 @@ class Agent:
         :param search: Looks for items with the specified string.
         :return: Dictionary: {'items': array of items, 'totalItems': Number of items (without applying the limit)}
         """
+        def get_hash(directory):
+            filename = "{0}/{1}/merged.mg".format(common.shared_path, directory)
+
+            with open(filename, 'rb') as f:
+                hashing.update(f.read())
+
+            return hashing.hexdigest()
 
         # Connect DB
         db_global = glob(common.database_path_global)
         if not db_global:
             raise WazuhException(1600)
+
+        # check hash algorithm
+        if not hash_algorithm in hashlib.algorithms_available:
+            raise WazuhException(1723)
+
+        hashing = hashlib.new(hash_algorithm) 
 
         conn = Connection(db_global[0])
         query = "SELECT {0} FROM agent WHERE `group` = :group_id"
@@ -1208,18 +1267,20 @@ class Agent:
         # Group names
         data = []
         for entry in listdir(common.shared_path):
-            item = {}
-
             full_entry = path.join(common.shared_path, entry)
             if not path.isdir(full_entry):
                 continue
 
-            item['name'] = entry
-
             # Group count
-            request = {'group_id': item['name']}
+            request = {'group_id': entry}
             conn.execute(query.format('COUNT(*)'), request)
-            item['count'] = conn.fetch()[0]
+
+            # merged.mg and agent.conf sum
+            merged_sum = get_hash(entry)
+            conf_sum   = get_hash(entry)
+
+            item = {'count':conn.fetch()[0], 'name': entry,
+                    'merged_sum': merged_sum, 'conf_sum': conf_sum}
 
             data.append(item)
 
@@ -1242,6 +1303,9 @@ class Agent:
         :param group_id: Group ID.
         :return: True if group exists, False otherwise
         """
+        # Input Validation of group_id
+        if not InputValidator().group(group_id):
+            raise WazuhException(1722)
 
         db_global = glob(common.database_path_global)
         if not db_global:
@@ -1269,6 +1333,9 @@ class Agent:
         :param group_id: Group ID.
         :return: True if group exists, False otherwise
         """
+        # Input Validation of group_id
+        if not InputValidator().group(group_id):
+            raise WazuhException(1722)
 
         if path.exists("{0}/{1}".format(common.shared_path, group_id)):
             return True
@@ -1276,7 +1343,7 @@ class Agent:
             return False
 
     @staticmethod
-    def get_agent_group(group_id, offset=0, limit=common.database_limit, sort=None, search=None):
+    def get_agent_group(group_id, offset=0, limit=common.database_limit, sort=None, search=None, select=None):
         """
         Gets the agents in a group
 
@@ -1294,12 +1361,24 @@ class Agent:
             raise WazuhException(1600)
 
         conn = Connection(db_global[0])
+        valid_select_fiels = ["id", "name", "ip", "last_keepalive", "os_name", 
+                             "os_version", "os_platform", "version",
+                             "config_sum", "merged_sum"]
 
         # Init query
         query = "SELECT {0} FROM agent WHERE `group` = :group_id"
         fields = {'id': 'id', 'name': 'name'}  # field: db_column
-        select = ['id', 'name']
         request = {'group_id': group_id}
+
+        # Select
+        if select:
+            if not set(select['fields']).issubset(valid_select_fiels):
+                uncorrect_fields = map(lambda x: str(x), set(select['fields']) - set(valid_select_fiels))
+                raise WazuhException(1724, "Allowed select fields: {0}. Fields {1}".\
+                        format(valid_select_fiels, uncorrect_fields))
+            select_fields = select['fields']
+        else:
+            select_fields = valid_select_fiels
 
         # Search
         if search:
@@ -1314,10 +1393,11 @@ class Agent:
         # Sorting
         if sort:
             if sort['fields']:
-                allowed_sort_fields = fields.keys()
+                allowed_sort_fields = select_fields
                 # Check if every element in sort['fields'] is in allowed_sort_fields.
                 if not set(sort['fields']).issubset(allowed_sort_fields):
-                    raise WazuhException(1403, 'Allowed sort fields: {0}. Fields: {1}'.format(allowed_sort_fields, sort['fields']))
+                    raise WazuhException(1403, 'Allowed sort fields: {0}. Fields: {1}'.\
+                        format(allowed_sort_fields, sort['fields']))
 
                 order_str_fields = ['{0} {1}'.format(fields[i], sort['order']) for i in sort['fields']]
                 query += ' ORDER BY ' + ','.join(order_str_fields)
@@ -1333,19 +1413,10 @@ class Agent:
             request['limit'] = limit
 
         # Data query
-        conn.execute(query.format(','.join(select)), request)
+        conn.execute(query.format(','.join(select_fields)), request)
 
-        data['items'] = []
-
-        for tuple in conn:
-            data_tuple = {}
-
-            if tuple[0] != None:
-                data_tuple['id'] = str(tuple[0]).zfill(3)
-            if tuple[1] != None:
-                data_tuple['name'] = tuple[1]
-
-            data['items'].append(data_tuple)
+        data['items'] = [{field:str(tuple_elem).zfill(3) for field,tuple_elem \
+                        in zip(select_fields, tuple) if tuple_elem} for tuple in conn]
 
         return data
 
@@ -1403,6 +1474,9 @@ class Agent:
         :param group_id: Group ID.
         :return: Confirmation message.
         """
+        # Input Validation of group_id
+        if not InputValidator().group(group_id):
+            raise WazuhException(1722)
 
         group_path = "{0}/{1}".format(common.shared_path, group_id)
 
@@ -1434,6 +1508,10 @@ class Agent:
         :return: Confirmation message.
         """
 
+        # Input Validation of group_id
+        if not InputValidator().group(group_id):
+            raise WazuhException(1722)
+
         if group_id.lower() == "default":
             raise WazuhException(1712)
 
@@ -1441,22 +1519,27 @@ class Agent:
             raise WazuhException(1710, group_id)
 
         ids = []
+        if isinstance(group_id, list):
+            for id in group_id:
+                try:
+                    Agent()._remove_single_group(id)
+                except Exception as e:
+                    ids.append(create_exception_dic(id, e))
+        else:
+            try:
+                Agent()._remove_single_group(group_id)
+            except Exception as e:
+                ids.append(create_exception_dic(group_id, e))
 
-        # Remove agent group
-        agents = Agent.get_agent_group(group_id=group_id, limit=None)
-        for agent in agents['items']:
-            Agent.unset_group(agent['id'])
-            ids.append(agent['id'])
+        final_dict = {}
+        if not ids:
+            message = 'All selected groups were removed'
+            final_dict = {'msg': message}
+        else:
+            message = 'Some groups were not removed'
+            final_dict = {'msg': message, 'ids': ids}
 
-        # Remove group directory
-        group_path = "{0}/{1}".format(common.shared_path, group_id)
-        group_backup = "{0}/groups/{1}_{2}".format(common.backup_path, group_id, int(time()))
-        if path.exists(group_path):
-            move(group_path, group_backup)
-
-        msg = "Group '{0}' removed.".format(group_id)
-
-        return {'msg': msg, 'affected_agents': ids}
+        return final_dict
 
     @staticmethod
     def set_group(agent_id, group_id, force=False):
@@ -1468,6 +1551,9 @@ class Agent:
         :param force: No check if agent exists
         :return: Confirmation message.
         """
+        # Input Validation of group_id
+        if not InputValidator().group(group_id):
+            raise WazuhException(1722)
 
         agent_id = agent_id.zfill(3)
         if agent_id == "000":
@@ -1514,7 +1600,6 @@ class Agent:
         :param force: No check if agent exists
         :return: Confirmation message.
         """
-
         # Check if agent exists
         if not force:
             Agent(agent_id).get_basic_information()
@@ -1763,7 +1848,7 @@ class Agent:
         s.close()
         if debug:
             print("RESPONSE: {0}".format(data))
-        if data.startswith('err'):
+        if data != 'ok':
             raise WazuhException(1715, data.replace("err ",""))
 
         # Sending file to agent
@@ -1774,7 +1859,7 @@ class Agent:
             print("Sending: {0}".format(common.ossec_path + "/var/upgrade/" + wpk_file))
         try:
             start_time = time()
-            bytes_read = file.read(512)
+            bytes_read = file.read(common.wpk_read_size)
             bytes_read_acum = 0
             while bytes_read:
                 s = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
@@ -1783,9 +1868,9 @@ class Agent:
                 s.send(msg.encode() + bytes_read)
                 data = s.recv(1024).decode()
                 s.close()
-                if data.startswith('err'):
+                if data != 'ok':
                     raise WazuhException(1715, data.replace("err ",""))
-                bytes_read = file.read(512)
+                bytes_read = file.read(common.wpk_read_size)
                 if show_progress:
                     bytes_read_acum = bytes_read_acum + len(bytes_read)
                     show_progress(int(bytes_read_acum * 100 / wpk_file_size) + (bytes_read_acum * 100 % wpk_file_size > 0))
@@ -1804,7 +1889,7 @@ class Agent:
         s.close()
         if debug:
             print("RESPONSE: {0}".format(data))
-        if data.startswith('err'):
+        if data != 'ok':
             raise WazuhException(1715, data.replace("err ",""))
 
         # Get file SHA1 from agent and compare
@@ -1818,6 +1903,8 @@ class Agent:
         s.close()
         if debug:
             print("RESPONSE: {0}".format(data))
+        if not data.startswith('ok '):
+            raise WazuhException(1715, data.replace("err ",""))
         rcv_sha1 = data.split(' ')[1]
         if rcv_sha1 == file_sha1:
             return ["WPK file sent", wpk_file]
@@ -1896,7 +1983,7 @@ class Agent:
         return Agent(agent_id).upgrade(wpk_repo=wpk_repo, version=version, force=True if int(force)==1 else False)
 
 
-    def upgrade_result(self, debug=False, timeout=60):
+    def upgrade_result(self, debug=False, timeout=common.upgrade_result_retries):
         """
         Read upgrade result output from agent.
         """
@@ -1914,7 +2001,7 @@ class Agent:
             print("RESPONSE: {0}".format(data))
         counter = 0
         while data.startswith('err') and counter < timeout:
-            sleep(1)
+            sleep(common.upgrade_result_sleep)
             counter = counter + 1
             s = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
             s.connect(common.ossec_path + "/queue/ossec/request")
@@ -1975,7 +2062,7 @@ class Agent:
         s.close()
         if debug:
             print("RESPONSE: {0}".format(data))
-        if data.startswith('err'):
+        if data != 'ok':
             raise WazuhException(1715, data.replace("err ",""))
 
         # Sending file to agent
@@ -1984,7 +2071,7 @@ class Agent:
             raise WazuhException(1715, data.replace("err ",""))
         try:
             start_time = time()
-            bytes_read = file.read(512)
+            bytes_read = file.read(common.wpk_read_size)
             file_sha1=sha1(bytes_read)
             bytes_read_acum = 0
             while bytes_read:
@@ -1994,7 +2081,7 @@ class Agent:
                 s.send(msg.encode() + bytes_read)
                 data = s.recv(1024).decode()
                 s.close()
-                bytes_read = file.read(512)
+                bytes_read = file.read(common.wpk_read_size)
                 file_sha1.update(bytes_read)
                 if show_progress:
                     bytes_read_acum = bytes_read_acum + len(bytes_read)
@@ -2017,7 +2104,7 @@ class Agent:
         s.close()
         if debug:
             print("RESPONSE: {0}".format(data))
-        if data.startswith('err'):
+        if data != 'ok':
             raise WazuhException(1715, data.replace("err ",""))
 
         # Get file SHA1 from agent and compare
@@ -2031,6 +2118,8 @@ class Agent:
         s.close()
         if debug:
             print("RESPONSE: {0}".format(data))
+        if not data.startswith('ok '):
+            raise WazuhException(1715, data.replace("err ",""))
         rcv_sha1 = data.split(' ')[1]
         if calc_sha1 == rcv_sha1:
             return ["WPK file sent", wpk_file]